Debugger: fix output formatting for addresses that appeared as 32-bit values

+++ src/CpuOpCodes.h
@@ -22,7 +22,7 @@ struct CpuOp {
     const char* name;
     AddressingMode addrMode;
     int cycles;
-    int size;
+    uint8_t size;
     const char* description;
 };
